test(scripts): split SC-008 into its own file to satisfy FR-006 ratchet
check-skill-assertions.test.ts is pre-existing over-cap (538 lines), and
the issue-#2069 severity assertions grew it to 539. FR-006's diff-scoped
ratchet fails when an over-cap file grows, which broke the quality job.

Moves the SC-008 drift-check-entry-point describe block into
check-skill-assertions-entrypoint.test.ts (172 lines), leaving the
original at 443 — both now under the 500-line cap. Test-repo helpers are
duplicated in the new file, matching the existing convention across the
other six check-skill-assertions test files.

No assertion semantics changed; 31 tests still pass, isolated and co-run.
